weaviate_gradio_local/requirements.txt
gradio==3.41.2
langchain==0.0.278
huggingface_hub==0.16.4
pypdf==3.15.4
transformers==4.32.1
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
sentence_transformers==2.2.2
weaviate-client==3.23.2
einops==0.6.1
accelerate==0.22.0
bitsandbytes==0.41.0